[feat] implement getLicensesAPI to LicenseListTab
- seperate Table to LicenseListTab
  - implement Table using useFetch hook
- change API route
