+++ src/components/containers/LicenseListTab/index.tsx
@@ -1,43 +1,28 @@
+// Dependencies
 import React, { useState } from 'react';
 import TabTemplate from '../../templates/TabTemplate';
-import Icon from '../../widgets/Icon';
-import Table from '../../widgets/Table';
+import AsyncBoundary from '../../../libs/AsyncBoundary';
+import * as Style from './styled';
+
+// Widgets
+import Table from './Table';
 import Input from '../../widgets/Input';
 import TabForm from '../../widgets/TabForm';
 import Dropdown from '../../widgets/Dropdown';
+
+// Modals
 import AddLicenseModal from '../../modals/AddLicenseModal';
 import DeleteModal from '../../modals/DeleteModal';
-import { licenseListAttr } from '../../../common/constants';
-import { LicenseListAttr, Number, Trash } from '../../../@types/types';
-import * as Style from './styled';
-
-export type ItemType = {
-  [key in LicenseListAttr]: string;
-};
+import Error from '../../widgets/Error';
 
 type ModalType = 'add' | 'delete' | 'none';
 
-export interface RowType extends ItemType {
-  [Number]: number;
-  [Trash]: React.ReactElement;
-}
-
-// FIXME: remove
-interface ProjectDetailTab {
-  items: ItemType[];
-}
-
-function LicenseListTab({ items }: ProjectDetailTab) {
+function LicenseListTab() {
   const [modalState, setModalState] = useState<ModalType>('none');
 
   const closeModal = () => setModalState('none');
   const openAddLicenseModal = () => setModalState('add');
   const openDeleteModal = () => setModalState('delete');
-  const parsedItem: RowType[] = items.map((item, index) => ({
-    ...item,
-    number: index + 1,
-    trash: <Icon onClick={openDeleteModal} icon="trashcan" size="2rem" />,
-  }));
 
   return (
     <>
@@ -51,9 +36,11 @@ function LicenseListTab({ items }: ProjectDetailTab) {
             <Dropdown label="규제" width="25rem" items={[]} onClickItem={() => {}} />
           </Style.InputWrapper>
         </TabForm>
-        <Style.TableWrapper>
-          <Table title="라이선스 목록" attributes={licenseListAttr} items={parsedItem} />
-        </Style.TableWrapper>
+        <AsyncBoundary pendingFallback={<>loading</>} rejectedFallback={Error}>
+          <Style.TableWrapper>
+            <Table openDeleteModal={openDeleteModal} />
+          </Style.TableWrapper>
+        </AsyncBoundary>
       </TabTemplate>
     </>
   );
